// src/app/pages/pago/interfaces/pago.interface.ts
export interface IPago{
    idpago?: string
    fecha: string
    idproducto?: string
    idconsignacion: string
    codconsignacion?: string
    codproducto?: string
    idcategoria?: string
    categoria?: string
    descripcion?: string
    talle?: string
    precio?: string
    activo?: string
    listCategoriaId?: string
}